# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: WeavePress API
  version: 0.1.0
  description: 内部内容采集、阅读和用户管理 API。
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: 存活检查
      responses: { "200": { description: OK } }
  /ready:
    get:
      summary: 依赖就绪检查
      responses: { "200": { description: Ready }, "503": { $ref: "#/components/responses/Error" } }
  /api/auth/login:
    post:
      summary: 登录
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginInput" } } }
      responses: { "200": { description: 登录成功, content: { application/json: { schema: { $ref: "#/components/schemas/TokenEnvelope" } } } }, "401": { $ref: "#/components/responses/Error" } }
  /api/auth/refresh:
    post:
      summary: 轮换 Refresh Token
      responses: { "200": { description: 刷新成功, content: { application/json: { schema: { $ref: "#/components/schemas/TokenEnvelope" } } } }, "401": { $ref: "#/components/responses/Error" } }
  /api/auth/logout:
    post:
      summary: 退出并撤销 Refresh Token
      responses: { "200": { description: 已退出 } }
  /api/user/info:
    get:
      summary: 当前用户
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK }, "401": { $ref: "#/components/responses/Error" } }
  /api/auth/codes:
    get:
      summary: 当前权限码
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK } }
  /api/menu/all:
    get:
      summary: 后端菜单兼容接口
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK } }
  /api/dashboard:
    get:
      summary: 工作台统计
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK } }
  /api/users:
    get:
      summary: 用户列表
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }, { name: keyword, in: query, schema: { type: string } }, { name: status, in: query, schema: { type: string, enum: [active, disabled] } } ]
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Error" } }
    post:
      summary: 创建用户
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateUserInput" } } }
      responses: { "201": { description: Created }, "403": { $ref: "#/components/responses/Error" } }
  /api/users/{id}:
    parameters: [ { $ref: "#/components/parameters/ID" } ]
    patch:
      summary: 修改用户
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK } }
    put:
      summary: 修改用户（管理后台兼容）
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK } }
  /api/users/{id}/password:
    parameters: [ { $ref: "#/components/parameters/ID" } ]
    put:
      summary: 重置密码并撤销登录会话
      security: [ { bearerAuth: [] } ]
      responses: { "200": { description: OK } }
  /api/collection-jobs:
    get:
      summary: 采集任务列表
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }, { name: status, in: query, schema: { type: string } } ]
      responses: { "200": { description: OK } }
    post:
      summary: 提交文章链接
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [url], properties: { url: { type: string, format: uri } } } } }
      responses: { "202": { description: 已进入队列 }, "400": { $ref: "#/components/responses/Error" } }
  /api/collection-jobs/{id}:
    get:
      summary: 任务详情和事件
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/ID" } ]
      responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/Error" } }
  /api/collection-jobs/{id}/retry:
    post:
      summary: 手动重试失败任务
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/ID" } ]
      responses: { "200": { description: 已重新排队 }, "409": { $ref: "#/components/responses/Error" } }
  /api/articles:
    get:
      summary: 文章列表
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }, { name: keyword, in: query, schema: { type: string } }, { name: sourceType, in: query, schema: { type: string, enum: [wechat, web] } }, { name: status, in: query, schema: { type: string, enum: [pending, processing, ready, failed] } } ]
      responses: { "200": { description: OK } }
  /api/articles/{id}:
    get:
      summary: 文章结构化详情
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/ID" } ]
      responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/Error" } }
  /api/articles/{id}/raw-url:
    get:
      summary: 原始快照短期签名地址
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/ID" } ]
      responses: { "200": { description: OK } }
  /media/assets/{id}:
    get:
      summary: 签名素材访问
      parameters:
        - { $ref: "#/components/parameters/ID" }
        - { name: expires, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: signature, in: query, required: true, schema: { type: string } }
      responses: { "200": { description: 本地文件 }, "307": { description: 重定向到七牛私有下载地址 }, "403": { description: 签名无效或过期 } }
  /media/raw/{id}:
    get:
      summary: 签名原始快照下载
      parameters:
        - { $ref: "#/components/parameters/ID" }
        - { name: expires, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: signature, in: query, required: true, schema: { type: string } }
      responses: { "200": { description: 本地 gzip 快照 }, "307": { description: 重定向到七牛私有下载地址 }, "403": { description: 签名无效或过期 } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    ID: { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  schemas:
    LoginInput:
      type: object
      required: [username, password]
      properties: { username: { type: string, minLength: 3 }, password: { type: string, minLength: 8 } }
    CreateUserInput:
      type: object
      required: [username, password, realName, role]
      properties: { username: { type: string }, password: { type: string, minLength: 8 }, realName: { type: string }, role: { type: string, enum: [admin, editor] } }
    TokenEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data:
          type: object
          required: [accessToken, expiresIn]
          properties: { accessToken: { type: string }, expiresIn: { type: integer, example: 900 } }
    Error:
      type: object
      required: [code, message]
      properties: { code: { type: integer }, message: { type: string }, details: { type: array, items: { type: object } } }
  responses:
    Error:
      description: 业务错误
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }